bot/requirements.txt
pyTelegramBotAPI
urllib3
python-dotenv
Psycopg2-binary